lxc-to-lxd: Fix handling on snap
Running lxc-to-lxd from a snap environment is tricky.

We can't just use the previous approach of resolving the main container
dir through HostPath as any included file will cause a config read
failure. We also can't just have the snap run the entire binary on the
host as the linker would fail on liblxc.

Instead, this approach lets the loader take care of the dynamic
libraries and we then attach to the host mount namespace and process the
rest from there.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>