lxd/images: Fix incorrect RBAC on push
An issue was introduced when image push support was introduced.

A "trusted" variable was introduced and the endpoint was marked as
allowing untrusted client but at the same time it was sitll limited to
an RBAC permission.

Until the recent cleanup, the RBAC wrapper would have just returned true
and moved on, but now it instead checks for a UserAccess struct which is
going to be missing in this case, then rightfully treats it as not
having the required permission and blocks it.

The fix is to remove this double check and correct the trusted check in
the function itself.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>