lxd/network/driver: Always delete when requested, ignore LocalStatus() pending
This is now up to caller to decide whether it is appropriate to delete or not (using n.LocalStatus()).

This is needed because if the network fails to start as part of a create operation on a local node we need to delete it, however it will still be in pending state and so would have previously been ignored.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>