lxd/instance/operationlock: Store operation in instanceOperations before calling go routine
As the go routine can call functions on the operation (such as op.Done) which rely on the instanceOperations map being populated it seems appropriate to ensure it has been populated with the new operation before starting the go routine.

Even though the only current use of the operation inside the go routine is after 30s.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>