lxd/device: Fix instance type validations
When validating the instance type as part of device config validation,
we need to keep in mind that instancetype.Any is a valid value.

This is used for profiles as they can apply to any of the types.

In such cases, the config validation should only error out when directly
applied to an instance and further checks should be added for a start
time error on instances getting the device through profiles.

Closes #8377

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>