forkproxy: prevent zombies
When I originally introduced forkproxy we pretty quickly ran into a zombie
problem because of interactions between the C "runtime" and the Go runtime,
quote:

> Usually we should wait for the child process somewhere here.
> But we cannot really do this. The listener file descriptors
> are retrieved in the go runtime but at that point we have
> already double-fork()ed to daemonize ourselves and so we
> can't wait on the child anymore after we received the
> listener fds. On the other hand, if we wait on the child
> here we wait on the child before the receive. However, if we
> do this then we can end up in a situation where the socket
> send buffer is full and we need to retrieve some file
> descriptors first before we can go on sending more. But this
> won't be possible because we're waiting before the call to
> receive the file descriptor in the go runtime. Luckily, we
> can just rely on init doing it's job and reaping the zombie
> process. So, technically unsatisfying but pragmatically
> correct.

Removing the double-fork in 28f7f6dc08d2 ("lxd/forkproxy: Switch to using
subprocess") meant that the child process didn't get reparented to init which
means it doesn't get cleaned up and hangs around as a zombie.

There are multiple solutions to this:
1. Ignore SIGCHLD in some form e.g. via signal() or sigaction().
2. Create a go routine in the parent process that waits.
3. Create a detached pthread() that waits.

Solution number 1. likely won't work because I expected Go to mess with signals
once the runtime kicks in; especially with SIGCHLD. Even if it didn't I
wouldn't want to rely on it not changing in the future.
Solution number 2. is doable but means more code and we'd need to transfer the
child PID to the go runtime which is fine but seems conceptually odd since the
process got created in pure C.
So solutioon 3. it is.

Link: https://discuss.linuxcontainers.org/t/forkproxy-child-process-hangs-after-adding-proxy-device-to-a-container
Fixes: 28f7f6dc08d2 ("lxd/forkproxy: Switch to using subprocess")
Cc: stable-4.0
Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>