lxd/instance/qemu: Add stateful stop/start
This currently comes with two regressions to security features.
 - No chroot
 - No privilege dropping

Both of those are incompatible with restoring VM state as unfortunately
QEMU performs privilege dropping prior to handling the migration stream
and so fails to then open the various files needed by the VM.

This could be worked around if QEMU would allow named FDs to be used to
back a drive, this way we could pass open FDs from the privileged side,
but this doesn't appear currently possible.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>