lxd/device/nic/ipvlan: Detach ipvlan interface back to random host name on stop, then delete
The kernel sometimes leaves references to ipvlan interfaces when the container's network namespace is destroyed.
This prevents recreation of the ipvlan interface with the same IP.

So instead of relying on the kernel to destroy the ipvlan interface on container stop, instead detach the ipvlan interface back to the host as a random name and then delete it.

Fixes #8427

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>