utils: trim whitespace from block device UUID
When creating a storage (btrfs) pool backed by a block device as in this
example

lxc storage create pool-btrfs btrfs source=/dev/sdc

we create the storage pool based on the volatile device path /dev/sdc. Since
this is obviously a bad idea we have logic to replace this device path with the
reliable /dev/disk/by-uuid/<uuid> path based on the uuid we just generated when
creating the filesystem. So we call

blkid -s UUID -o value /dev/sdc

and use the returned uuid in a tryExists(/dev/disk/by-uuid/<uuid>) loop for 10
seconds. This consistently failed for me which means I ended up with the
unreliable /dev/sdc device path as the source property of the storage pool
which I couldn't change later. The reason this happens is that whitespace
is sneaking in making the path whose existence we test for useless. Fix this by
trimming whitespace.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>