lxd/network/acl/acl/ovn: Adds OVNPortGroupDeleteIfUnused
Reworks PortGroupDeleteIfUnused from network package to detect all unused port groups.

As searching for 1 unused ACL port group requires searching through all instance NICs, profile NICs, networks and ACL rules, it is not much more expensive to search for the status of all ACLs at the same time (no more DB queries, just a bit more CPU work search configs for multiple ACL names).

It is also required to do this wide search now that ACL rules can refer to other ACLs (which themselves may not be directly in use by an OVN entity).

As such we now gather usage information for all ACLs and split it by direct-OVN-referred usage and indirect-ACL-rule-refererred usage.

For the direct-OVN-referred in-use ACLs, we do not delete them.
For indirect-ACL-rule-refererred in-use ACLs, we collect the names of the ACLs that are referring to other ACLs, and check if they are in directly in-use by an OVN entity. If they are not it is safe to remove them as well.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>