lxd/db: Don't fail preparing statements for activateifneeded
On startup, we call "lxd activateifneeded", this does a number of simple
database access using sqlite directly. We only need a fraction of the
tables at that point and this will run ahead of any on-startup database
schema update.

This change makes it so that new prepared statements for tables that do
not yet exist won't prevent activateifneeded from accessing the
database.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>