lxd/db/images: Fix incorrect cached attribute handling
This fixed a very nasty regression introduced in 9c1bc785d8 which
switched the function fetching expired images to filter on a per-project
basis.

That part made total sense but unfortunately what we missed is that we
had no code-generated DB code for that particular filter
(Project-Cached) so instead only the project filter was used, causing
the Cached field to be ignored entirely.

This then resulted in all expired images to be returned to the image
pruner EVEN if an image wasn't cached, causing the deletion of manually
created images that weren't used recently...

Reported at: https://discuss.linuxcontainers.org/t/images-vanished/10498/

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>