cluster: Larger Timeout to find leader
go-dqlite has an internal timeout of 5s when attempting to contact a
node. If the timeout of the parent context is not large enough, an
unreachable node might consume the whole time budget, and other,
possible reachable nodes will not be attempted anymore leading to
observed failures while in fact the cluster is healthy.

Signed-off-by: Mathieu Borderé <mathieu.bordere@canonical.com>