lxd/storage/drivers/driver/zfs/volumes: Only mount instance filesystem volumes in postHook for CreateVolumeFromBackup
Otherwise calling MountVolume for an instance block volume will increment the filesystem volume mount ref counter again, as MountVolume would have already been called for the filesystem volume as part of the recursive call to CreateVolumeFromBackup for VM Block volumes above.

And as we don't use the post hook function from that function call, we'd end up leaving the filesystem volume of a block volume mounted due to the extra ref count.

Fixes #8649

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>