lxd/firewwall/drivers/drivers/nftables: Changes nftables to use a single inet table rather than separate ip and ip6 tables
The rationale behind this is to simplify the forthcoming ACL rules that will be added as a single chain with mixed IP4/IPv6 rules, rather than needing to separate them over two tables.

This also makes it easier to reason about where LXD is putting its rules, and can result in rule amount reduction by being able to use the same rule for both IPv4 and IPv6 traffic.

I've ensure that the cleanup functions still remove rules from the old ip and ip6 tables, as well as the new inet table, to ensure that old rules get removed for running instances and existing networks.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>