lxd/patches: Adds patchClusteringServerCertTrust
This patch adds the local server cert to the global trusted server certificates list.

It then waits for the rest of the cluster members to do the same.

Finally it refreshes the local cached trusted certificates list and switches to using the local server certificate as the intra-cluster client certificate.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>