lxd: use idmapped mounts
Add support for using idmapped mounts when starting new containers.
Add a new function that tests and caches the information whether the
container's rootfs supports idmapped mounts or not.

If the kernel supports idmapped mounts for the container's filesystem
type LXD will now make use of them. If not we will fallback to shiftfs
if it is supported on this system.

We introduce new types allowing is to discern between no-idmapped
storage support, idmapped storage support and shiftfs support.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>