lxd/instance/drivers/driver/qemu: Converts NICs to be added via QMP rather than static config
This is a precursor to supporting hot pluggable NICs, but first making the static NICs effectively hotplugged at start up.

To do this, pci ports are still allocated via static config, but then a QMP monitor hook function is returned to be run after the qemu process has been started (and thus QMP connected), but before the emulation has begun.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>