lxd/cluster/gateway: Only create persistent warning for cluster time skew when cluster is operational
Switches to d.cluster.UpsertWarningLocalNode and warnings.ResolveWarningsByLocalNodeAndType usage.

Previously this was preventing the heartbeat request from being handled normally if the cluster DB was fully functioning as the query to retrieve the local name failed.

Also cleans up the logging to log the actual error, and avoids the DB query to look up the name in normal situations and only when a skew is detected or fixed do we need to look up the name.

We still log the warning to syslog though as that doesn't require DB access.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>