lxd/storage/drivers/utils: Adds force arg to shrinkFileSystem
If true will pass -f flag to the resize2fs command to disable safety checks to allow shrinking when resize2fs makes an incorrect estimate of how much space is needed to complete the shrink.

This is required, at least for LVM, in certain circumstances, as the pre-resize checks are not always accurate and can prevent successful resize.

This is inspired by the fsadm.sh script in the lvm2 distribution, which calls resize2fs with the `-f` flag when using the `-r` flag of lvresize with the `-f` flag (as we do in LXD).

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>