lxd/storage/drivers/driver/zfs/volumes: Fix bug with VM optimized import not returning filesystem volume post hook
This was causing the filesystem volume to be left mounted and the mount counter was not decremented after import.
Which could then later causes issue if volume needed to be unmounted.

Unlike other storage drivers, the ZFS driver was recursively calling CreateVolumeFromBackup to handle the VM filesystem volume.
This commit changes that so it handles the VM filesystem volume directly.

This way we can then generate only a single post hook and also the revert in the post hook is correctly handled too (which was also missing before).

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>