lxd/device/nic: Return -1 for Mtu in State() for bridged and ovn NICs if host interface not available
Rather than failing the entire State() call.

This ensures that if a client is polling the /state endpoint to find when a VM has stopped that it won't sometimes return an error during the VM shutdown process, if the NIC has been cleaned up but the VM is not yet recorded as stopped.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>