client: Only retry target addresses if initial connection fails
Don't try next address if the initial operation creation succeeds, but then the actual operation fails.

This suggests that the error is internal to the server and retrying the same operation just on a different address for the same target can potentially exacerbate the problem.

In one example #8900 the instance copy to a remote target node failed on the target due to insufficent pool size.
This error was successfully being returned to the client, but it subsequently retried, but the original operation was not cleaned up, causing a hang.

There was no reason to retry as this could not have succeeded on a different address.

Fixes #8900

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>