shared/instance: Add evacuation options
This add two new config options `cluster.evacuate` and
`volatile.evacuate.origin`.

The `cluster.evacuate' option defines the action to be taken
when performing the evacuation. It takes one of the following values:

* auto (default, migrate the instance if possible, otherwise stop)
* stop (stop the instance)
* migrate (migrate the instance)

The value of `volatile.evacuate.origin` is the node where the instances
originate from. On recovery, all migrated instances will be migrated
back to this node.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>