lxd/device: Fix duplicate MAC test
The current check would require BOTH the parent and network properties
to differ for the MAC check to be skipped. This is effectively
impossible and was clearly meant to be an OR instead.

Additionally, if two interfaces share the same parent but are on
different VLANs, they also should skip the MAC duplication check.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>