lxd/instance/lxc: Rework raw.lxc handling
We need to make sure that raw.lxc is applied last in the instance's config.
This is quite hard to do because of devices which are mostly dynamic and
added after the initial configuration is generated.

So instead, this creates a separate configuration file which is always
loaded following the main file. To keep validation working properly,
this then moves the validation logic into Update().

With this change, `raw.lxc` is only applied when a container is started,
the other operations should not need that configuration.

Closes #9093

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>