shared/api/error: Adds StatusError type
For use in both LXD server and in the LXD client for passing an HTTP status code in an error to the caller.

This can be used in the LXD server to alter the HTTP status code returned to clients if an error occurs,
and in the LXD client it can be used to propagate the HTTP status code returned from the server to the user
so they can handle errors by inspecting the status code rather than inspecting the error message itself
which may change over time.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>