lxd/endpoints/network: don't give up if no network listeners exist
This change is to address a bug where if there is no network listener, we
do not fall back to cluster listener. This can happen if
"cluster.https_address" is specified but "core.https_address" is not.
This change allows us to ignore the network listener and move on to
check for a cluster listener, rather than just returning prematurely.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>