lxd/db/generate/db/mapping: add TableType to Mapping
The goal of TableType is to split up the generator’s behavior based on
the kinds of fields in the struct it’s using. The struct’s fields should
map to table columns, so TableType determines whether a table exists as
an independent entity, or is referenced by another table (via
ReferenceID field), is representing a map (key/value fields), etc.

What I’m trying to do with all this is make it so each table that we
have can be represented by a struct that the generator can then parse
and discern some context from, so the appropriate methods can be
generated.

This means that when we generate methods for instances for example, we
don’t need to assume the structure of completely different reference
tables for devices, profiles, etc.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>