seccomp: verify retrieved fds when
Pidfds currently can only be used with thread-group leaders.  When we
get a request from a non-threadgroup leader to perform a bpf syscall we
parse out the threadgroup leader's pid and allocate a pidfd based on the
threadgroup leader's pid. We then proceed to retrieve file descriptors
based on the threadgroup leader's pidfd. This assumes that the
threadgroup leader and the target thread share the same file descriptor
table but that assumption isn't safe because the target thread could
have called unshare(CLONE_FILES) at some point, closed the bpf fds and
opened new ones. So we need to make sure that the retrieved fds that we
now hold refer to the same open file as the target thread. To achieve
this we use kcmp(). If kcmp() tells us that the file descriptor table is
still shared after we retrieved the fds we're good. After this we also
need to make sure that the process the pidfd refers to is still alive to
make sure we're talking about the fds of the same threadgroup.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>