lxd/cluster: Fix forceful stop on restore
This fixes a bug where if an instance wouldn't stop within its deadline,
we would return nil at that point and so silently abort the restore
entirely.

This was because the old logic would check for
    "err != nil || errors.Cause(err) != drivers.ErrInstanceIsStopped"
and so a return value of "nil" be treated as satisfying the condition
and be returned causing the restore operation to succeed despite not
having completed nor having cleared the flag from the database.

Took me a little while to track that down as the lack of any error was
quite disconcerting :)

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>