lxd/instance_exec: Improve and expand default PATH setting
This moves the PATH checks to hit /proc instead of using the expensive
FileExists function. This will potentially fail if the path being
checked is an absolute symlink or requires going through an absolute
symlink, but those paths are expected to be directories so this should
be fine.

This also adds the NixOS path (/run/current-system/sw/bin) alongside the
existing snap path (/snap/bin).

Closes #9467

Reported-by: Maciej Krüger <mkg20001@gmail.com>
Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>