lxd/api/cluster: Directly disable clustering when removing self
When a node requests to remove itself from the database, the request is
forwarded to the leader, which removes the node from dqlite and then
sends another request to the to-be-removed node to disable clustering on
itself and re-exec its daemon. This made it so that flushing the
response in clusterPutDelete was rendering to the leader node, and the
daemon on the requesting node would restart before the response could be
sent back.

Instead of asking the leader to request the to-be-removed node to
disable clustering, this commit makes the original requesting node
to immediately disable clustering on itself and re-exec its daemon if
the request goes through.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>