Allow USB pass through of specific port path
With the fixes to USB device pass-through, allowing multiple identical
devices to be used inside one virtual machine, users may want to route
specific USB devices to specific instances. This can be done by
following the physical topology of hubs and ports.

The new `portglob` property can be used to optionally match the port
path to a glob. Port path is a sequence of numbers separated with
dashes, with each number representing a port and dash representing
connections between hubs.

This allows adding arbitrary device plugged to a physical port or a
particular hub to a specific instance. The property is supported for
`usb` device type both virtual machines and on containers and for
`unix-hotplug` on containers only.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>