doc: Add initial sphinx support
This adds support for building our documentation into "doc/html" using Sphinx.

It's designed to be as unintrusive as possible, effectively keeping
everything within a hidden ".sphinx" directory with as few duplicated
code/dependencies.

A new `make doc` target is added to fetch the needed dependencies inside
of a python virtual environment, download the external theme and build
everything.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>
Inspired-by: Ruth Fuchss <ruth.fuchss@canonical.com>