lxd/instance/exec: Rework Connect to make control connection required for interactive exec
But not required for non-interactive exec.
Due to a historical bug in the LXC CLI command, we cannot force the client to connect a control socket when in non-interactive mode.
This is because the older CLI tools did not connect this channel and so we would prevent the older CLIs connecitng to newer servers.
So skip the control connection from being considered as a required connection in this case.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>