lxd/instance/exec: Adds a check for remote stdout reader closing and kills command if no control connection
We don't expect to read anything from the remote side of the stdout websocket.
If we do then this is a protocol exception, and could indicate the remote side has disconnected (EOF).

If the control connection hasn't connected then we kill the command to ensure that if the client disconnects
prematurely we don't leave processes running in the instance.

Fixes #9578

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>