lxd/storage/drivers/driver/btrfs/volumes: Enable nodatacow on block raw disk files
Enable nodatacow so that random writes don't cause fragmentation and old extents to be kept.
BTRFS extents are immutable so when blocks are written they end up in new extents and the old
ones remains until all of its data is dereferenced or rewritten. These old extents are counted
in the quota, and so leaving CoW enabled can cause the BTRFS subvolume quota to be reached even
if the block volume file isn't full.

Fixes #9124

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>