lxd-user: Initial implementation
This introduces a new command `lxd-user` which acts as a proxy to the
main LXD API. It listens on its own unix socket and handles:

 - Initial configuration of the LXD daemon, creating a default storage
   pool, default network and default profile on first connection.
 - On initial connection from a user, it will generate a new certificate
   keypair, it then generates a dedicated project for the user and ties
   that certificate to the project.
 - The user interacts with that LXD normally through the CLI or API but
   will only see their own project and instances.

This effectively allows running both the `lxd` and `lxd-user` daemon on
a system. The main `lxd` daemon's socket is restricted to those who
administer the entire system (members of the `lxd` group usually)
whereas the `lxd-user` daemon's socket is accessible to all local users
(`users` group or similar).

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>