lxd/db/images: Updates imageFillProfiles to set image.Profiles=nil when project has its own profiles but not its own images
If a project's features.profiles=true but its features.images=false, then ensure that image.Profiles is set to nil to allow
instance.CreateInternal() to detect this and fallback to the "default" profile in the project for the image.

Otherwise instance creation will fail if profiles are not specified at instance creation time.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>