lxd/cluster/membership: Renames notifyNodesUpdate to NotifyHeartbeat
 - Improved logging.
 - Updates to use NewAPIHeartbeat (properly initialising heartbeat).
 - Updates to send all member info in notification heartbeat.

It was concerning me that this function was using the Raft ID as the Member ID in the heartbeat data, as they are not guaranteed
to be the same. Although hopefully the recipient wasn't using it, we shouldn't rely on that!

 - Reads the raft nodes from the local DB (as we expect any updates to have been written by then).

This will make it easier to call this function when updating cluster member roles.

 - Adds offlineThreshold argument to NotifyHeartbeat
 - Changes the heartbeat type to full state. This way it will be able to trigger member refresh tasks on the recipients.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>