lxd/db: Use ClusterTx with ToAPI for filling reference fields
As we don't necessarily want to query every table associated with an
entity each time, pass in a transaction to ToAPI so that the fields can
at least be filled for the public facing representations.

Additionally, InstanceArgs in db/instances.go will be used for sharing
information across instances and snapshots that would normally be in
their API structs. Hence, db.InstanceToArgs now expects an api.Instance,
and the new method snapshot.ToInstanceArgs will populate reference
fields with a ClusterTx argument.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>