lxd/db: Updates GetInstanceNamesByNodeAddress to also return project name.
This function is called on GET /1.0/instances to get a list of instance
names when querying for more/expanded data. However, instance names are
not globally unique and this leads to incorrect aggregation of instances
when "all-projects=true". This change returns the project name as well
as the instance name so that aggregations are performed correctly.

Additionally, the GetInstanceToNodeMap function is altered to return a
map in which the keys are arrays containing both the project and
instance name. This would have led to incorrect reporting of nodes when
listing instances from all projects if clustered.

Signed-off-by: Mark Laing <mark.laing@canonical.com>