lxd/images: Don't cleanup unknown images from shared volume
We have logic which deletes all unknown image files on startup, but this
assumes that the image storage volume isn't shared with other servers.

As we now support using a cephfs volume for shared image storage, we're
finding ourselves deleting images downloaded by other servers on startup
which isn't exactly ideal.

This adds a reasonably fast check to see if we're using a shared storage
volume and if so, skip that logic entirely.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>