lxc/list: Add InstanceFull shortcut
When `lxc list` is unable to just use `?recursion=2`, it falls back to
pulling the state and snapshot information directly from the instance.

But when one needs both the state and the snapshots, it now makes more
sense to pull ?recursion=1 and get half of the API queries.

On a server supporting ?recursion=1 this leads to an identical response
in about half the time.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>