lxd/db: Fix volume sequence regression
The previous schema update effectively recreated most of the LXD schema.

While doing so, it accidentally interfered with the weird sequence we
have shared between storage_volumes and storage_volumes_snapshots.

Following the schema update, that sequence ends up set to the maximum
value of the storage_volumes table when it should be set to the maximum
value of either storage_volumes or storage_volumes_snapshots whichever
is highest.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>