doc: fix broken links
Linking to the automatic ID does not work in Sphinx when appending
the automatic ID to the md file name.
It works within a page because you don't need the file name, but
this is error prone (because the ID depends on the heading and
changes when the heading changes) and should be changed as well in
the future.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>