doc: enable automatically generated anchors
Enable
https://myst-parser.readthedocs.io/en/latest/syntax/optional.html#auto-generated-header-anchors
Not sure if this really is a great improvement, since it adds more
options for linking, which will make it harder to be consistent.
It does allow for simpler links using Markdown syntax, but we can
only use those when overriding the link text (because otherwise,
they won't show up on GitHub at all).
When sticking to MyST syntax, linking gets a bit more complicated,
but we don't have to add manual targets for every link.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>