lxd/db: Fix stoarge_volumes sequence (again)
The previous queries would lead to incorrect results if one of the two
tables was empty, usually resulting in a sequence of 0 rather than the
max of the storages_volumes table.

This now uses a simpler query using one sub-query per table making it
far easier to understand and validate.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>