lxd/metrics: Rework caching and locking
This moves all caching and locking to the metrics API file.

It also changes the cache to work on a per-project basis and for the get
endpoint to similarly retrieve the data per project.

This avoids cache corruption when fetching for an individual project and
makes it so we only ever retrieve data from projects that haven't been
fetched in the past 15s.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>