lxc/console: Rework concurency model for vga
This now properly handles counting connections, disconnecting when
they're all close or on server error or on operation completion.

It also ensures that no channels can be closed twice and that we can't
be reading from a closed channel which should avoid some of the panics
we've recently seen in that logic.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>