lxd/storage/drivers/driver/zfs: Set all dataset mountpoint settings to legacy
This is in relation to https://github.com/lxc/lxd-pkg-snap/issues/61 and trying to avoid
the ZFS tooling running inside the snap mount namspace from trying to mount/unmount the
datasets as it can get confused and fail to find the mounts.

Our experimentation in earier iterations of these changes suggest that using the legacy
mount/umount commands fair better than getting the ZFS tooling do it.

We previously set mountpoint=none but have seen that this is causing ZFS tooling to
try and unmount volumes which is still causing problems.

This is an attempt to try and further avoid ZFS from trying to mount/unmount datasets.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>