lxd/storage/drivers/btrfs: Fix migration handling
Newer version of btrfs (tested with v5.15.1) fail to do migration
the way it's currently done. The error message comes from the receiver
and is `ERROR: cannot find parent subvolume`.

This changes the receiving end to place the received stream in the
correct location (no temporary directory) without having to first mark
the snapshot read-write and then move it around (which causes the
issue).

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>