lxd/db/generate/db/method/v2: Add V2 method generation
To simplify the process of refactoring a large chunk of the generator,
this introduces a "version" flag that can be passed to the generator to
pick between 2 different method generation implementations. Currently
this is being used for the 'certificate_projects' table methods. The
main change is to pass in and return actual db entity structs into these
methods so that we can easily use them outside of other generated
methods (as they are currently being used).

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>