lxd/storage/load: Update GetPoolByInstance to use instance's StoragePool() function
This ensures that it will use an pre-loaded storage pool var inside the instance,
to avoid extra DB queries.

This is also essential as we move to creating the storage volume DB records in the
storage package, as at instance create time the storage pool to use is derived from
the instance's root disk config and then that storage pool is pre-loaded inside the
instance. Meaning that we can then access it to create the instance's volume, whereas
without this the storage volume select query would fail as the DB record hasn't been
created yet.

Also improves error messages.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>