lxd/instance/drivers/qemu: Fix block devices
Drives were always using the "file" driver which required the block
device file to be a regular file. The "file" driver however doesn't work
with character or or block devices.

This fixes the issue by checking whether the file is a character or
block device. In this case, the "host_device" driver is used.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>